Update README.md
This project was based on 2 case studies where concepts of Linear Regression & Logistic Regression were used and respective models were build in order for prediction.
In 1st case study, concepts of linear regression were used to build a model to predict the price of gems. In 2nd case study, concepts of Logistic Regression were used to build a model so that it could predict that customer will purchase the holiday package or not.